Registro delete terminados

+++ src/auditorias/auditorias.controller.ts
@@ -1,18 +1,23 @@
 import { Controller, Post, UseGuards, Body, Get, Query } from "@nestjs/common";
 import { AuthGuard } from "../guards/auth.guard";
-
 import {
   AuditoriasInterface,
   AuditoriaRespuestas,
 } from "./auditorias.interface";
 import { AuditoriasService } from "./auditorias.class";
 import { Tienda } from "src/tiendas/tiendas.class";
+import { LoggerService } from "src/logger/logger.service";
+import { UserRecord } from "firebase-admin/auth";
+import { User } from "../decorators/get-user.decorator";
+import { TrabajadorService } from "src/trabajadores/trabajadores.class";
 
 @Controller("auditorias")
 export class AuditoriasController {
   constructor(
     private readonly auditoriaInstance: AuditoriasService,
     private readonly tiendasInstance: Tienda,
+    private readonly loggerService: LoggerService,
+    private readonly trabajadores: TrabajadorService,
   ) {}
 
   @UseGuards(AuthGuard)
@@ -189,16 +194,40 @@ export class AuditoriasController {
   //Borrar auditoria
   @UseGuards(AuthGuard)
   @Post("deleteAuditoria")
-  async deleteAuditoria(@Body() auditoria: AuditoriasInterface) {
+  async deleteAuditoria(
+    @Body() auditoria: AuditoriasInterface,
+    @User() user: UserRecord,
+  ) {
     try {
+      const auditoriaToDelete = await this.auditoriaInstance.getAuditoriasById(
+        auditoria,
+      );
+      if (!auditoriaToDelete) {
+        throw new Error("Auditoria no encontrada");
+      }
+
       const respAuditoria = await this.auditoriaInstance.deleteAuditoria(
         auditoria,
       );
-      if (respAuditoria)
+
+      if (respAuditoria) {
+        // Obtener el nombre del usuario autenticado
+        const usuarioCompleto = await this.trabajadores.getTrabajadorByAppId(
+          user.uid,
+        );
+        const nombreUsuario = usuarioCompleto?.nombreApellidos || user.email;
+        // Registro de la auditoría
+        await this.loggerService.create({
+          action: "Eliminar Auditoria",
+          name: nombreUsuario,
+          extraData: { auditoriaData: auditoriaToDelete },
+        });
+
         return {
           ok: true,
           data: respAuditoria,
         };
+      }
 
       throw Error("No se ha podido borrar la auditoria");
     } catch (err) {
